+++ src/common/io.ts
@@ -51,6 +51,7 @@ import {
   AoEpochSettings,
   AoGateway,
   AoGatewayDelegateWithAddress,
+  AoGatewayRegistrySettings,
   AoGatewayVault,
   AoIORead,
   AoIOWrite,
@@ -454,6 +455,14 @@ export class IOReadable implements AoIORead {
     quantity: number;
     name: string;
   }): Promise<number>;
+  async getTokenCost(params: {
+    intent: 'Upgrade-Name';
+    name: string;
+  }): Promise<number>;
+  async getTokenCost(params: {
+    intent: 'Primary-Name-Request';
+    name: string;
+  }): Promise<number>;
   async getTokenCost({
     intent,
     type,
@@ -461,10 +470,15 @@ export class IOReadable implements AoIORead {
     name,
     quantity,
   }: {
-    intent: 'Buy-Record' | 'Extend-Lease' | 'Increase-Undername-Limit';
+    intent:
+      | 'Buy-Record'
+      | 'Extend-Lease'
+      | 'Increase-Undername-Limit'
+      | 'Upgrade-Name'
+      | 'Primary-Name-Request';
     type?: 'permabuy' | 'lease';
     years?: number;
-    name?: string;
+    name: string;
     quantity?: number;
   }): Promise<number> {
     const allTags = [
@@ -710,6 +724,12 @@ export class IOReadable implements AoIORead {
         { name: 'Action', value: 'Redelegation-Fee' },
         { name: 'Address', value: params.address },
       ],
+    });
+  }
+
+  async getGatewayRegistrySettings(): Promise<AoGatewayRegistrySettings> {
+    return this.process.read({
+      tags: [{ name: 'Action', value: 'Gateway-Registry-Settings' }],
     });
   }
 }